refactor: 의존성 역전을 통한 순환참조 제거
