+++ src/main/java/kitchenpos/order/application/OrderMapper.java
@@ -6,7 +6,6 @@ import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderTableRepository;
 import kitchenpos.order.domain.OrderValidator;
 import kitchenpos.order.domain.OrderedMenu;
 import kitchenpos.order.dto.request.OrderLineItemRequest;
@@ -16,13 +15,10 @@ import org.springframework.stereotype.Component;
 @Component
 public class OrderMapper {
 
-    private final OrderTableRepository orderTableRepository;
     private final MenuRepository menuRepository;
     private final OrderValidator orderValidator;
 
-    public OrderMapper(final OrderTableRepository orderTableRepository, final MenuRepository menuRepository,
-                       final OrderValidator orderValidator) {
-        this.orderTableRepository = orderTableRepository;
+    public OrderMapper(final MenuRepository menuRepository, final OrderValidator orderValidator) {
         this.menuRepository = menuRepository;
         this.orderValidator = orderValidator;
     }
