+++ src/main/java/kitchenpos/table/application/TableGroupMapper.java
@@ -2,8 +2,8 @@ package kitchenpos.table.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.order.domain.OrderTableRepository;
 import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table.domain.TableGroup;
 import kitchenpos.table.dto.request.TableGroupRequest;
 import kitchenpos.table.dto.request.TableGroupRequest.TableId;
