+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -2,10 +2,10 @@ package kitchenpos.table.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.order.domain.OrderTableRepository;
 import kitchenpos.order.dto.request.OrderTableRequest;
 import kitchenpos.order.dto.response.OrderTableResponse;
 import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table.domain.OrderTableValidator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
