+++ src/main/java/kitchenpos/table/domain/OrderTableRepository.java
@@ -1,8 +1,7 @@
-package kitchenpos.order.domain;
+package kitchenpos.table.domain;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.table.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
