+++ src/test/java/kitchenpos/table/application/TableGroupServiceTest.java
@@ -13,7 +13,6 @@ import kitchenpos.menu.dto.request.MenuRequest;
 import kitchenpos.menu.dto.response.MenuGroupResponse;
 import kitchenpos.menu.dto.response.MenuResponse;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTableRepository;
 import kitchenpos.order.dto.request.OrderLineItemRequest;
 import kitchenpos.order.dto.request.OrderRequest;
 import kitchenpos.order.dto.request.OrderStatusUpdateRequest;
@@ -23,6 +22,7 @@ import kitchenpos.order.dto.response.OrderTableResponse;
 import kitchenpos.product.domain.Product;
 import kitchenpos.support.ServiceTest;
 import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table.domain.TableGroupRepository;
 import kitchenpos.table.dto.request.TableGroupRequest;
 import kitchenpos.table.dto.response.TableGroupResponse;
